requirements.txt
# Core LangChain - Vertex AI Setup
langchain>=0.1.20
langchain-community>=0.0.38
langchain-google-vertexai>=1.0.0

# Google Cloud Platform - Canadian Compliance Stack
google-cloud-bigquery>=3.14.1
google-cloud-aiplatform>=1.40.0
google-cloud-secret-manager>=2.17.0
google-cloud-logging>=3.9.0

# Database and data processing
pandas>=2.1.4
pyarrow>=14.0.2
db-dtypes>=1.2.0

# Security and encryption - CRITICAL for mental health data
cryptography>=41.0.7
python-dotenv>=1.0.0
pyjwt>=2.8.0

# API and web framework (optional - for REST API)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Logging and monitoring - HIPAA compliance
structlog>=24.1.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0
httpx>=0.26.0

# Code quality tools (development only)
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.2
bandit>=1.7.6

# Type stubs
types-requests>=2.31.0.20240106